pyproject.toml
[project]
name = 'Project'
description = 'Project Description'
authors = [
    {name = 'Cam Ratchford', email = 'camratchford@gmail.com'},
]

classifiers = [
    'Development Status :: 1 - Testing required',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3 :: Only',
    'Intended Audience :: Developers',
    'Intended Audience :: Information Technology',
    'Intended Audience :: System Administrators',
    'Operating System :: Windows :: Windows',
    'Environment :: Console',
    'Environment :: Windows',
    'Environment :: MacOS X',
    'Topic :: Software Development :: Libraries :: Python Modules',
    'Topic :: Internet',
]
version = "1.0.0"
requires-python = '>=3.7'
dependencies = [
    "aiosqlite==0.19.0",
    "annotated-types==0.6.0",
    "anyio==3.7.1",
    "asyncpg==0.28.0",
    "bcrypt==4.0.1",
    "certifi==2023.7.22",
    "cffi==1.16.0",
    "click==8.1.7",
    "colorama==0.4.6",
    "cryptography==41.0.5",
    "dnspython==2.4.2",
    "email-validator==2.0.0.post2",
    "exceptiongroup==1.1.3",
    "fastapi==0.104.0",
    "FastAPI-SQLAlchemy==0.2.1",
    "fastapi-users==12.1.2",
    "fastapi-users-db-sqlalchemy==6.0.1",
    "greenlet==3.0.1",
    "h11==0.14.0",
    "httpcore==0.17.3",
    "httptools==0.6.1",
    "httpx==0.24.1",
    "httpx-oauth==0.13.0",
    "idna==3.4",
    "makefun==1.15.1",
    "markdown-it-py==3.0.0",
    "mdurl==0.1.2",
    "passlib==1.7.4",
    "pycparser==2.21",
    "pydantic==2.4.2",
    "pydantic_core==2.10.1",
    "Pygments==2.16.1",
    "PyJWT==2.8.0",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.6",
    "PyYAML==6.0.1",
    "rich==13.6.0",
    "sniffio==1.3.0",
    "SQLAlchemy==2.0.22",
    "starlette==0.27.0",
    "typing_extensions==4.8.0",
    "uvicorn==0.23.2",
    "watchfiles==0.21.0",
    "websockets==12.0",
]

[project.urls]
Homepage = 'https://github.com/camratchford/Project'
Source = 'https://github.com/camratchford/Project'

[project.scripts]
app_run = "app.cli:run"